--- lib/index.ts
@@ -1,6 +1,0 @@
-export { Box } from "./Box";
-export { Buttons } from "./Buttons";
-export { HeroSection } from "./HeroSection";
-export { Modal } from "./Modal";
-export { Notifications } from "./Notifications";
-export { TableOfContents } from "./TableOfContents";
