--- lib/types/deps.d.ts
@@ -1,4 +1,0 @@
-// reactstrap.d.ts
-declare module 'https://esm.sh/reactstrap' {
-  export * from 'reactstrap'
-}
