+++ src/stories/HeroSection.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import "bulma/css/bulma.min.css";
-import { HeroSection } from "../../lib";
+import { HeroSection } from "../components/HeroSection";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
@@ -22,9 +22,11 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Basic: Story = {
   args: {},
-  render: () => <HeroSection>
+  render: () => (
+    <HeroSection>
       <h1 className="title">Title</h1>
       <h2 className="subtitle">Subtitle</h2>
       <button className="button is-warning">Learn More</button>
-  </HeroSection>,
+    </HeroSection>
+  ),
 };
